% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disastLocations_df.R
\docType{data}
\name{disastLocations_df}
\alias{disastLocations_df}
\title{disastLocations_df}
\format{
A tibble with
\describe{
\item{eventKey}{The ID created by the authors to inform the year,
state, and number of disasters in that particular place}
\item{state}{Abbreviation of the state where an event happened}
\item{county}{Name of the county the disaster}
}
}
\usage{
data(disastLocations_df)
}
\description{
This data contains the the states and counties where disasters
have happened across the USA.
}
\keyword{datasets}
