% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disastTypes_df.R
\docType{data}
\name{disastTypes_df}
\alias{disastTypes_df}
\title{disastTypes_df}
\format{
A tibble with
\describe{
\item{eventKey}{The ID created by the authors to inform the year,
state, and number of disasters in that particular place}
\item{incident_type}{It is the original type of disaster that the FEMA
and EMDAT data sets reported}
\item{hazard_type}{It is the related broad classification disaster type
that was assigned by the authors according to the original incident type
reported by FEMA or EMDAT. The category was assigned based on the
Technical report}
\item{hazard_cluster}{It is the related sub type classification assigned by
the authors according to the original incident type reported by FEMA or
EMDAT, and acciording to the broad classification that was already
assigned}
}
}
\usage{
data(disastTypes_df)
}
\description{
This data contains contains the type of disasters reported
by FEMA and EMDAT based on the classification  of the Hazard Definition and
Classification Review Technical Report  published by the UN Office for
Disaster Risk Reduction (2020)
\url{https://www.undrr.org/publication/hazard-definition-and-classification-review}
}
\keyword{datasets}
