% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emdat_hazard_cluster_df.R
\docType{data}
\name{emdat_hazard_cluster_df}
\alias{emdat_hazard_cluster_df}
\title{emdat_hazard_cluster_df}
\format{
A tibble with
\describe{
\item{Source}{This vector will contain the information of the data set
from which the information came from, in this case EMDAT}
\item{incident_type}{It is the original type of disaster that the EMDAT
data sets reports}
\item{hazard_cluster}{It is the related sub type classification assigned by
the authors according to the original incident type reported by EMDAT
and acciording to the broad classification that was already assigned}
}
}
\usage{
data(fema_hazard_cluster_df)
}
\description{
This data contains the type of disasters reported
by EMDAT  and the matched hazard cluster based on the classification  of
the Hazard Definition and  Classification Review Technical Report
published by the UN Office for  Disaster Risk Reduction (2020)
Annex 6 page 72. The marching was done by Dr. Mark Macgowan and
Catalina Canizares following the suggestions in the report. However,
this classification is subjective and should be used taking into
account this disclaimer.
\url{https://www.undrr.org/publication/hazard-definition-and-classification-review}
}
\keyword{datasets}
